Version 1.0 - Initial commit
WDToolkit is a collection of functionalities a webdeveloper might want to use.
It is capable of encoding in different styles, hashing strings and files, doing bit operations
and finally it comes with a color picker, allowing the user to chose a color from a picker which
is then displayed as color itself, hex and rgb values.